requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.12.3
chardet==3.0.4
conda==4.3.16
cryptography==2.7
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
django-rest-knox
drf-spectacular==0.17.3
enum-compat==0.0.3
gunicorn==20.1.0
idna==2.8
importlib-metadata==4.6.3
jsonschema==3.2.0
libarchive-c==2.8
psycopg2==2.8.5
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.0.0
pyrsistent==0.18.0
PySocks==1.7.0
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
simplejson==3.17.0
six==1.12.0
sqlparse==0.4.1
tqdm==4.32.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.24.2
whitenoise==5.3.0
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.5.0
